requirements.txt
fastapi~=0.109.0
mysql-connector-python~=8.2.0
python-decouple~=3.8
uvicorn[standard]~=0.25.0
